change usage to Inst and remove last_local
